question/0106 Construct Binary Tree from Inorder and Postorder Traversal_test.go
package question

import (
	"reflect"
	"testing"
)

func Test_buildTree106(t *testing.T) {
	type args struct {
		inorder   []int
		postorder []int
	}
	tests := []struct {
		name string
		args args
		want *TreeNode
	}{
		{
			name: "Test1",
			args: args{
				inorder:   []int{9, 3, 15, 20, 7},
				postorder: []int{9, 15, 7, 20, 3},
			},
			want: makeTree([]int{3,9,20,-1,-1,15,7},-1),
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := buildTree106(tt.args.inorder, tt.args.postorder); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("buildTree106() = %v, want %v", got, tt.want)
			}
		})
	}
}
